Resolved a linker error where modulus was called on a double, and a bug where the wrong weapon was called for Crazy Random Sword.

+++ CrazyRandomSword.cpp
@@ -3,7 +3,7 @@
 double CrazyRandomSword::hit(double armor) {
 	srand(time(NULL));
 	hitPoints = rand() % 90 + 10;
-	double armIgnored = rand() % ((armor/2) + 1);
+	double armIgnored = static_cast<int>(rand()) % static_cast<int>(((armor/2) + 1));
 	double damage = hitPoints - (armor - armIgnored);
     if (damage < 0) {
         return 0;
